requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosmtpd==1.4.6
alembic==1.14.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
arrow==1.3.0
asgiref==3.8.1
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
atpublic==5.1
attrs==24.3.0
binaryornot==0.4.4
build==1.2.2.post1
certifi==2024.8.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.9.0
contourpy==1.3.1
cookiecutter==2.6.0
cycler==0.12.1
debugpy==1.8.12
decorator==4.4.2
distro==1.9.0
Django==5.0.9
django-appconf==1.0.6
django-browser-reload==1.17.0
django-compressor==4.5.1
django-cors-headers==4.6.0
django-decouple==2.1
django-sass-processor==1.4.1
django-tailwind==3.8.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
drf-spectacular==0.28.0
einops==0.8.0
exceptiongroup==1.2.2
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
HyperPyYAML==1.2.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflection==0.5.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
kiwisolver==1.4.8
libsass==0.23.0
lightning==2.5.0.post0
lightning-utilities==0.12.0
llvmlite==0.44.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.6.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
mysqlclient==2.2.6
networkx==3.4.2
noisereduce==3.0.3
nose==1.3.7
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.59.6
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
optuna==4.2.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip-tools==7.4.1
primePy==1.3
proglog==0.1.11
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-slugify==8.0.4
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2025.1
PyYAML==6.0.2
rcssmin==1.1.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rjsmin==1.2.2
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.1
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.2
SQLAlchemy==2.0.38
sqlparse==0.5.2
srt==3.5.3
stripe==8.9.0
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.8.0
tomli==2.2.1
torch==2.6.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.6.0
torchmetrics==1.6.1
tqdm==4.67.1
triton==3.2.0
typer==0.15.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
vosk==0.3.44
websockets==14.1
webvtt-py==0.5.1
yarl==1.18.3
selenium==4.15.0
